#49 [mod] : questId 타입 Long으로 변경

+++ app/src/main/java/com/byeboo/app/presentation/quest/QuestReviewScreen.kt
@@ -46,7 +46,7 @@ import kotlinx.coroutines.flow.collectLatest
 
 @Composable
 fun QuestReviewScreen(
-    questId: Int,
+    questId: Long,
     navigateToQuest: () -> Unit,
     viewModel: QuestReviewViewModel = hiltViewModel()
 ) {
