+++ app/src/main/java/com/byeboo/app/presentation/quest/component/QuestBox.kt
@@ -8,10 +8,10 @@ import com.byeboo.app.presentation.quest.QuestState
 
 @Composable
 fun QuestBox(
-    questId: Int,
+    questId: Long,
     questNumber: Int,
     state: QuestState,
-    onQuestClick: (Int) -> Unit = {},
+    onQuestClick: (Long) -> Unit = {},
     modifier: Modifier = Modifier
 ) {
     val isClickable = state is QuestState.Available || state is QuestState.Complete
