+++ app/src/main/java/com/byeboo/app/presentation/quest/record/QuestRecordingViewModel.kt
@@ -34,7 +34,7 @@ class QuestRecordingViewModel @Inject constructor() : ViewModel() {
     val showQuitModal: StateFlow<Boolean>
         get() = _showQuitModal.asStateFlow()
 
-    fun setQuestId(questId: Int) {
+    fun setQuestId(questId: Long) {
         _state.update {
             it.copy(questId = questId)
         }
